產線頁面排版優化 產線回來的訂單查詢及分配(未完善) 頁簽問題已解決

+++ InternalSystem/Models/PersonnelLeaveOver.cs
@@ -10,8 +10,8 @@ namespace InternalSystem.Models
         public int EmployeeId { get; set; }
         public int LeaveType { get; set; }
         public int Quantity { get; set; }
-        public int Used { get; set; }
-        public int LeaveOver { get; set; }
+        public double Used { get; set; }
+        public double LeaveOver { get; set; }
 
         public virtual PersonnelProfileDetail Employee { get; set; }
         public virtual PersonnelLeaveType LeaveTypeNavigation { get; set; }
